backend/requirements.txt
google-generativeai
Flask
python-dotenv